Moved test to tracersdk
